Move SectionMissingBannerComponent under candidate namespace
