+++ app/components/candidate_interface/becoming_a_teacher_review_component.html.erb
@@ -1,5 +1,5 @@
 <% if show_missing_banner? %>
-  <%= render(SectionMissingBannerComponent.new(section: :becoming_a_teacher, section_path: candidate_interface_becoming_a_teacher_edit_path, error: @missing_error)) %>
+  <%= render(CandidateInterface::SectionMissingBannerComponent.new(section: :becoming_a_teacher, section_path: candidate_interface_becoming_a_teacher_edit_path, error: @missing_error)) %>
 <% else %>
   <%= render(SummaryCardComponent.new(rows: becoming_a_teacher_form_rows, editable: @editable)) %>
 <% end %>
