+++ app/components/candidate_interface/contact_details_review_component.html.erb
@@ -1,5 +1,5 @@
 <% if show_missing_banner? %>
-  <%= render(SectionMissingBannerComponent.new(section: :contact_details, section_path: candidate_interface_contact_details_edit_base_path, error: @missing_error)) %>
+  <%= render(CandidateInterface::SectionMissingBannerComponent.new(section: :contact_details, section_path: candidate_interface_contact_details_edit_base_path, error: @missing_error)) %>
 <% else %>
   <%= render(SummaryCardComponent.new(rows: contact_details_form_rows, editable: @editable)) %>
 <% end %>
