+++ app/components/candidate_interface/degrees_review_component.html.erb
@@ -13,5 +13,5 @@
 <% end %>
 
 <% if show_missing_banner? %>
-  <%= render(SectionMissingBannerComponent.new(section: :degrees, section_path: candidate_interface_new_degree_path, error: @missing_error)) %>
+  <%= render(CandidateInterface::SectionMissingBannerComponent.new(section: :degrees, section_path: candidate_interface_new_degree_path, error: @missing_error)) %>
 <% end %>
