+++ app/components/candidate_interface/gcse_qualification_review_component.html.erb
@@ -1,5 +1,5 @@
 <% if show_missing_banner? %>
-  <%= render(SectionMissingBannerComponent.new(section: "#{subject}_gcse", section_path: candidate_interface_gcse_details_edit_type_path(subject: subject.to_sym), error: @missing_error)) %>
+  <%= render(CandidateInterface::SectionMissingBannerComponent.new(section: "#{subject}_gcse", section_path: candidate_interface_gcse_details_edit_type_path(subject: subject.to_sym), error: @missing_error)) %>
 <% elsif @application_qualification %>
   <%= render(SummaryCardComponent.new(rows: gcse_qualification_rows, editable: @editable)) do %>
     <% title = if @application_qualification.missing_qualification?
