+++ app/components/candidate_interface/interview_preferences_review_component.html.erb
@@ -1,5 +1,5 @@
 <% if show_missing_banner? %>
-  <%= render(SectionMissingBannerComponent.new(section: :interview_preferences, section_path: candidate_interface_interview_preferences_edit_path, error: @missing_error)) %>
+  <%= render(CandidateInterface::SectionMissingBannerComponent.new(section: :interview_preferences, section_path: candidate_interface_interview_preferences_edit_path, error: @missing_error)) %>
 <% else %>
   <%= render(SummaryCardComponent.new(rows: interview_preferences_form_rows, editable: @editable)) %>
 <% end %>
