+++ app/components/candidate_interface/other_qualifications_review_component.html.erb
@@ -1,5 +1,5 @@
 <% if show_missing_banner? %>
-  <%= render(SectionMissingBannerComponent.new(section: :other_qualifications, section_path: candidate_interface_review_other_qualifications_path, error: @missing_error)) %>
+  <%= render(CandidateInterface::SectionMissingBannerComponent.new(section: :other_qualifications, section_path: candidate_interface_review_other_qualifications_path, error: @missing_error)) %>
 <% end %>
 <% @qualifications.each do |qualification| %>
   <%= render(SummaryCardComponent.new(rows: other_qualifications_rows(qualification), editable: @editable)) do %>
