+++ app/components/candidate_interface/personal_details_review_component.html.erb
@@ -1,5 +1,5 @@
 <% if show_missing_banner? %>
-  <%= render(SectionMissingBannerComponent.new(section: :personal_details, section_path: candidate_interface_personal_details_edit_path, error: @missing_error)) %>
+  <%= render(CandidateInterface::SectionMissingBannerComponent.new(section: :personal_details, section_path: candidate_interface_personal_details_edit_path, error: @missing_error)) %>
 <% else %>
   <%= render(SummaryCardComponent.new(rows: rows, editable: @editable)) %>
 <% end %>
