+++ app/components/candidate_interface/section_missing_banner_component.html.erb
