+++ app/components/candidate_interface/subject_knowledge_review_component.html.erb
@@ -1,5 +1,5 @@
 <% if show_missing_banner? %>
-  <%= render(SectionMissingBannerComponent.new(section: :subject_knowledge, section_path: candidate_interface_subject_knowledge_edit_path, error: @missing_error)) %>
+  <%= render(CandidateInterface::SectionMissingBannerComponent.new(section: :subject_knowledge, section_path: candidate_interface_subject_knowledge_edit_path, error: @missing_error)) %>
 <% else %>
   <%= render(SummaryCardComponent.new(rows: subject_knowledge_form_rows, editable: @editable)) %>
 <% end %>
