+++ app/components/candidate_interface/training_with_a_disability_review_component.html.erb
@@ -1,5 +1,5 @@
 <% if show_missing_banner? %>
-  <%= render(SectionMissingBannerComponent.new(section: :training_with_a_disability, section_path: candidate_interface_training_with_a_disability_edit_path, error: @missing_error)) %>
+  <%= render(CandidateInterface::SectionMissingBannerComponent.new(section: :training_with_a_disability, section_path: candidate_interface_training_with_a_disability_edit_path, error: @missing_error)) %>
 <% else %>
   <%= render(SummaryCardComponent.new(rows: training_with_a_disability_form_rows, editable: @editable)) %>
 <% end %>
