+++ spec/components/candidate_interface/section_missing_banner_component_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe SectionMissingBannerComponent do
+RSpec.describe CandidateInterface::SectionMissingBannerComponent do
   it 'renders a section missing banner component' do
     result = render_inline(described_class.new(section: 'degrees', section_path: '#'))
     expect(result.css('.app-banner__message .govuk-body').text).to include('Degree(s) section not marked as completed')
